auto_remaster/sandbox/labeling_tool/requirements.txt
fastapi
uvicorn
datasets
Pillow
aiofiles
